Move PPO1/TRPO seg_gen to commons
